requirements.txt
httpx
pyyaml
